+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 
 ### Internal Changes
 
+- [MR-4529](https://gitlab-master.nvidia.com/nvmetal/carbide/-/merge_requests/4529): chore: tests: Added: `TestManagedHost::machine_validation_completed`
+
 ## [v2025.08.29-rc2-0](https://gitlab-master.nvidia.com/nvmetal/carbide/-/compare/v2025.08.15-rc2-0...v2025.08.29-rc2-0)
 
 ### Added